backend/requirements.txt
fastapi
uvicorn
osmnx
networkx
shapely
scikit-learn
python-dotenv